requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.15
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
PyJWT==1.7.1
pytz==2024.1
sqlparse==0.5.1
typing-extensions==4.12.2
